Create temparature_converter.java
This program allows the user to enter a temperature value and select the conversion option. The available options are converting Celsius to Fahrenheit or Fahrenheit to Celsius. Based on the user's choice, the program calls the corresponding conversion method (celsiusToFahrenheit or fahrenheitToCelsius) to perform the temperature conversion.

The converted temperature is then displayed on the screen. The program handles invalid choices by displaying an error message.

You can run this program and test different temperature conversions by compiling and executing it.